resolve issues from request review
