+++ app/package.json
@@ -34,13 +34,13 @@
     ]
   },
   "devDependencies": {
-    "@types/react-copy-to-clipboard": "^4.3.0",
-    "@types/styled-components": "^5.1.2",
     "@types/node": "^14.0.27",
     "@types/react": "^16.9.44",
+    "@types/react-copy-to-clipboard": "^4.3.0",
     "@types/react-dom": "^16.9.8",
     "@types/react-map-gl": "^5.2.6",
     "@types/react-router-dom": "^5.1.5",
+    "@types/styled-components": "^5.1.2",
     "typescript": "^3.9.7"
   }
 }
