+++ app/src/components/pages/CafeDetail/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router";
-import { CafeInfo } from "../../sections/MainFeed";
 import { cafeApiURL, HomeScale } from "../Main";
 import CafeDetails from "../../sections/CafeDetails";
+import { CafeInfo } from "../../../utils";
 
 const CafeDetailPage = () => {
   const { cafeId }: { cafeId: string } = useParams();
