+++ app/src/components/sections/CafeImageSlide/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import SlideOne from "../ImageSlideOne";
 import ImageCarousel from "../ImageCarousel";
-import { CafeInfo } from "../MainFeed";
+import { CafeInfo } from "../../../utils";
 
 type CafeImageSlideProps = {
   cafe: CafeInfo | null;
