+++ app/src/components/sections/MainFeed/index.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 import PlacePreviewList from "../PlacePreviewList";
-import { Menus } from "../MenuCategory";
 import MainPost from "../MainPost";
 import "./index.css";
 import MaterialIcon from "../../common/MaterialIcon";
+import { CafeInfo } from "../../../utils";
 
 const MContainer = styled.div`
   display: flex;
@@ -27,22 +27,6 @@ const FeedBox = styled.div`
   align-items: stretch;
   flex: 0 0 360px;
 `;
-
-export type PlaceInfo = {
-  placeCategory: string;
-  cafes: CafeInfo[];
-};
-
-export type CafeInfo = {
-  id: string;
-  name: string;
-  place: string;
-  imageUris: string[];
-  mainImageUri: string;
-  americanoPrice: number;
-  floor: number;
-  menus: Menus;
-};
 
 type MainFeedProps = {
   cafeList: CafeInfo[] | null;
