+++ app/src/utils/index.tsx
@@ -1,4 +1,36 @@
 import styled from "styled-components";
+import { isMobile, isAndroid, isIOS } from "react-device-detect";
+
+export type Menus = {
+  categories: MenuCategory[] | null;
+};
+
+export type MenuCategory = {
+  categoryName: string;
+  categoryMenu: MenuPrice[];
+};
+
+export type MenuPrice = {
+  name: string;
+  ename: string;
+  price: string;
+};
+
+export type PlaceInfo = {
+  placeCategory: string;
+  cafes: CafeInfo[];
+};
+
+export type CafeInfo = {
+  id: string;
+  name: string;
+  place: string;
+  imageUris: string[];
+  mainImageUri: string;
+  americanoPrice: number;
+  floor: number;
+  menus: Menus;
+};
 
 export const DetailContainer = styled.div`
   display: flex;
@@ -58,50 +90,51 @@ export const copyLink = (cafeName: string | undefined) => {
 
 export const openSearch = (searchedData: string, searchEngine: string) => {
   const encodedData = encodeURIComponent(searchedData);
-  const userAgent = navigator.userAgent;
 
-  if (userAgent.match(/iPhone|iPad|iPod/i)) {
-    setTimeout(function () {
-      switch (searchEngine) {
-        case "Naver":
-          window
-            .open(
-              "https://m.search.naver.com/search.naver?query=" + encodedData,
-              "_blank"
-            )
-            ?.focus();
-          break;
+  if (isMobile) {
+    if (isIOS) {
+      setTimeout(function () {
+        switch (searchEngine) {
+          case "Naver":
+            window
+              .open(
+                "https://m.search.naver.com/search.naver?query=" + encodedData,
+                "_blank"
+              )
+              ?.focus();
+            break;
 
-        case "Instagram":
-          window
-            .open(
-              "https://www.instagram.com/explore/tags/" + encodedData,
-              "_blank"
-            )
-            ?.focus();
-          break;
-      }
-    }, 0);
-  } else if (userAgent.match(/android|Android/i)) {
-    setTimeout(function () {
-      switch (searchEngine) {
-        case "Naver":
-          window.location.href =
-            "intent://inappbrowser?url=https%3A%2F%2Fm.search.naver.com%2Fsearch.naver%3Fquery%3D" +
-            encodedData +
-            "&target=new&version=6#Intent;scheme=naversearchapp;action=android.intent.action.VIEW;category=android.intent.category.BROWSABLE;package=com.nhn.android.search;end";
-          break;
+          case "Instagram":
+            window
+              .open(
+                "https://www.instagram.com/explore/tags/" + encodedData,
+                "_blank"
+              )
+              ?.focus();
+            break;
+        }
+      }, 0);
+    } else if (isAndroid) {
+      setTimeout(function () {
+        switch (searchEngine) {
+          case "Naver":
+            window.location.href =
+              "intent://inappbrowser?url=https%3A%2F%2Fm.search.naver.com%2Fsearch.naver%3Fquery%3D" +
+              encodedData +
+              "&target=new&version=6#Intent;scheme=naversearchapp;action=android.intent.action.VIEW;category=android.intent.category.BROWSABLE;package=com.nhn.android.search;end";
+            break;
 
-        case "Instagram":
-          window
-            .open(
-              "https://www.instagram.com/explore/tags/" + encodedData,
-              "_blank"
-            )
-            ?.focus();
-          break;
-      }
-    }, 0);
+          case "Instagram":
+            window
+              .open(
+                "https://www.instagram.com/explore/tags/" + encodedData,
+                "_blank"
+              )
+              ?.focus();
+            break;
+        }
+      }, 0);
+    }
   } else {
     window.confirm("모바일 웹만 지원하는 서비스입니다.");
   }
